// frontend-mobile/api/generated/model/partidoResource.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Resultados River API Documentation
 * API for managing and querying historical football results for River Plate.
 * OpenAPI spec version: 1.0.0
 */
import type { ArbitroResource } from './arbitroResource';
import type { CondicionResource } from './condicionResource';
import type { EstadioResource } from './estadioResource';
import type { FaseResource } from './faseResource';
import type { GolResource } from './golResource';
import type { PartidoResourceResultado } from './partidoResourceResultado';
import type { RivalResource } from './rivalResource';
import type { TecnicoResource } from './tecnicoResource';
import type { TorneoResource } from './torneoResource';

export interface PartidoResource {
  fecha?: string;
  goles_river?: number;
  goles_rival?: number;
  /** @nullable */
  observaciones?: string | null;
  resultado?: PartidoResourceResultado;
  torneo?: TorneoResource;
  rival?: RivalResource;
  arbitro?: ArbitroResource;
  estadio?: EstadioResource;
  condicion?: CondicionResource;
  fase?: FaseResource;
  tecnico?: TecnicoResource;
  goles?: GolResource[];
}